+++ packages/core/src/clients/handlers/fetch.ts
@@ -8,7 +8,7 @@ import { withMemoization } from '../utils/memoization';
 import { AmplifyErrorCode } from '../../types';
 
 const shouldSendBody = (method: string) =>
-	!['HEAD', 'GET', 'DELETE'].includes(method.toUpperCase());
+	!['HEAD', 'GET'].includes(method.toUpperCase());
 
 // TODO[AllanZhengYP]: we need to provide isCanceledError utility
 export const fetchTransferHandler: TransferHandler<