+++ packages/auth/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.14.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.13.3...@aws-amplify/auth@6.14.0) (2025-07-23)
+
+### Features
+
+- refresh token rotation ([#14427](https://github.com/aws-amplify/amplify-js/issues/14427)) ([e5a8569](https://github.com/aws-amplify/amplify-js/commit/e5a85697101735c21548e397832fab12d001a5dc))
+
+## [6.13.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.13.2...@aws-amplify/auth@6.13.3) (2025-07-03)
+
+**Note:** Version bump only for package @aws-amplify/auth
+
+## [6.13.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.13.1...@aws-amplify/auth@6.13.2) (2025-07-02)
+
+### Bug Fixes
+
+- remove import side effects for getAmplifyUserAgent() ([#14433](https://github.com/aws-amplify/amplify-js/issues/14433)) ([f94a30b](https://github.com/aws-amplify/amplify-js/commit/f94a30b71e4e5abbce1ad8c215fbb67386721bff))
+
+## [6.13.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.13.0...@aws-amplify/auth@6.13.1) (2025-06-17)
+
+**Note:** Version bump only for package @aws-amplify/auth
+
+# [6.13.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.12.4...@aws-amplify/auth@6.13.0) (2025-05-27)
+
+### Features
+
+- **auth:** support custom identity pool endpoint ([#14382](https://github.com/aws-amplify/amplify-js/issues/14382)) ([fd56414](https://github.com/aws-amplify/amplify-js/commit/fd56414205509e383e635aceb29b9a780eed2dfc))
+
 ## [6.12.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/auth@6.12.3...@aws-amplify/auth@6.12.4) (2025-04-28)
 
 **Note:** Version bump only for package @aws-amplify/auth