+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.60](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.59...@aws-amplify/pubsub@6.1.60) (2025-07-23)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.59](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.58...@aws-amplify/pubsub@6.1.59) (2025-07-03)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.58](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.57...@aws-amplify/pubsub@6.1.58) (2025-07-02)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.57](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.56...@aws-amplify/pubsub@6.1.57) (2025-06-17)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.56](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.55...@aws-amplify/pubsub@6.1.56) (2025-05-27)
+
+### Bug Fixes
+
+- **pubsub:** fix ESM compatibility issue ([0087e4f](https://github.com/aws-amplify/amplify-js/commit/0087e4f26c11e1c8f059dd79d8acdb97205de0c5))
+
 ## [6.1.55](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.54...@aws-amplify/pubsub@6.1.55) (2025-04-28)
 
 **Note:** Version bump only for package @aws-amplify/pubsub