+++ scripts/tsc-compliance-test/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.90](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.89...tsc-compliance-test@0.1.90) (2025-07-23)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.89](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.88...tsc-compliance-test@0.1.89) (2025-07-03)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.88](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.87...tsc-compliance-test@0.1.88) (2025-07-02)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.87](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.86...tsc-compliance-test@0.1.87) (2025-06-17)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.86](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.85...tsc-compliance-test@0.1.86) (2025-05-27)
+
+**Note:** Version bump only for package tsc-compliance-test
+
 ## [0.1.85](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.84...tsc-compliance-test@0.1.85) (2025-04-28)
 
 **Note:** Version bump only for package tsc-compliance-test